pipeline/requirements.txt
# Core dependencies
requests==2.31.0
anthropic==0.39.0
python-dotenv==1.0.0

# Document processing
beautifulsoup4==4.12.3
lxml==5.1.0
PyMuPDF==1.23.8  # fitz - Better PDF parsing
spacy==3.7.2  # Sentence tokenization
tiktoken==0.5.2  # Token counting

# Vector database and embeddings
chromadb==0.4.22
sentence-transformers==2.2.2  # Text embeddings (includes cross-encoders)
rank-bm25==0.2.2  # BM25 keyword search

# Testing
pytest==8.0.0
pytest-mock==3.12.0

# Utilities
tenacity==8.2.3  # Retry logic
pydantic==2.5.3  # Data validation
tqdm==4.66.1  # Progress bars
cachetools==5.3.2  # Caching with TTL
